feat : 굿즈 상세 조회 기능

+++ src/main/java/com/efubtoy/team1/domain/goods/domian/Goods.java
@@ -21,9 +21,9 @@ public class Goods {
     @Column(name = "goods_name", length = 50)
     private String goodsName;
 
-    private int price;
+    private Long price;
 
-    private int stock;
+    private Long stock;
 
     @Column(length = 20)
     private String location;
